+++ makefile
@@ -2,7 +2,7 @@ cc = g++
 CFLAGS = -Wall -Wextra
 
 all:
-	$(cc) $(CFLAGS) -o build/main.out src/main.c src/physics.c src/shapes.c src/globalStructs.c src/globalFunctions.c src/loadShader.cpp -lGL -lglfw -lGLEW
+	$(cc) $(CFLAGS) -o build/main.out src/main.c src/physics.c src/shapes.c src/helperFiles/globalStructs.c src/helperFiles/globalFunctions.c src/shaders/loadShader.cpp -lGL -lglfw -lGLEW
 
 run: all
 	./build/main.out
