+++ src/helperFiles/globalFunctions.c
