+++ src/helperFiles/globalFunctions.h
