+++ src/helperFiles/globalStructs.c
