+++ src/helperFiles/globalStructs.h
